:white_check_mark: Feat: add React Essentials complete
