--- Gracekim/Section3/01-starting-project/src/components/TabButton.jsx
@@ -1,8 +1,0 @@
-export default function TabButton({children, onSelect }) {
-  console.log('TABBUTTON COMPONENT EXECUTING');
-  return (
-    <li>
-      <button onClick={onSelect}>{children}</button>
-    </li>
-  );
-}
