+++ Gracekim/Section3/React Essentials/src/App.jsx
@@ -1,0 +1,89 @@
+import { useState } from 'react';
+
+import { CORE_CONCEPTS } from './data.js';
+import Header from './components/Header/Header.jsx'
+import CoreConcept from './components/CoreConcept.jsx'
+import TabButton from './components/TabButton.jsx';
+import { EXAMPLES } from './data.js';
+
+
+function App() {
+  const [ selectedTopic, setSelectedTopic ] = useState();
+
+  function handleSelect(selectedButton) {
+    // selectedButton => 'components', 'jsx', 'props', 'state
+    setSelectedTopic(selectedButton);
+    // 로그를 출력했을 때 이전 값이 나옴.
+    console.log(selectedTopic);
+  }
+
+  console.log('APP COMPONENT EXECUTING');
+
+  let tabContent = <p>Please select a topic.</p>;
+
+  if (selectedTopic) {
+    tabContent = (
+      <div id="tab-content">
+            {selectedTopic} <p>Please select a topic.</p>
+            <h3>{EXAMPLES[selectedTopic].title}</h3>
+            <p>{EXAMPLES[selectedTopic].description}</p>
+            <pre>
+              <code>
+                {EXAMPLES[selectedTopic].code}
+              </code>
+            </pre>
+            
+          </div>
+    );
+  }
+
+  return (
+    <div>
+      <Header />
+      <main>
+        <section id="core-concepts">
+        <h2>Core Concepts</h2>
+        
+        <ul>
+          {CORE_CONCEPTS.map((conceptItem) => (
+            <CoreConcept key={conceptItem.title} {...conceptItem} />
+          ))}
+          
+        </ul>
+        </section>
+        <section id="examples">
+          <h2>Examples</h2>
+          <menu>
+            <TabButton
+              isSelected={selectedTopic === 'components'}
+              onSelect={() => handleSelect('components')}
+             >
+              Component
+             </TabButton>
+             <TabButton
+              isSelected={selectedTopic === 'jsx'}
+              onSelect={() => handleSelect('jsx')}
+             >
+              JSX
+             </TabButton>
+             <TabButton
+              isSelected={selectedTopic === 'props'}
+              onSelect={() => handleSelect('props')}
+             >
+              Props
+             </TabButton>
+             <TabButton
+              isSelected={selectedTopic === 'state'}
+              onSelect={() => handleSelect('state')}
+             >
+              State
+             </TabButton>
+          </menu>
+          {tabContent}
+        </section>
+      </main>
+    </div>
+  );
+}
+
+export default App;
