+++ Gracekim/Section3/React Essentials/src/assets/react-core-concepts.png
