+++ Gracekim/Section3/React Essentials/src/components/CoreConcept.jsx
