+++ Gracekim/Section3/React Essentials/src/components/TabButton.jsx
@@ -1,0 +1,8 @@
+export default function TabButton({children, onSelect, isSelected }) {
+  console.log('TABBUTTON COMPONENT EXECUTING');
+  return (
+    <li>
+      <button className={isSelected ? 'active': undefined} onClick={onSelect}>{children}</button>
+    </li>
+  );
+}
